requirements.txt
python-dotenv
farcaster
gspread
oauth2client
